enh: Move profile from `/` to `/profile`
