+++ frontend/src/routes/profile/+page.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
 	import Navbar from '$lib/components/Navbar.svelte';
-	import type { PageData } from './$types';
-	export let data: PageData;
+	export let data;
 </script>
 
 <Navbar />
 
 <!-- TODO: Get name and profile from server.ts => providerUser to db -->
 <h1>Profile</h1>
-<h2>User id: {data.user.userId}</h2>
+<h2>User ID: {data.user.userId}</h2>
 <h2>Email: {data.user.email}</h2>
 <h2>Name: {data.user.name}</h2>
